backend/requirements.txt
APScheduler==3.6.3
asgiref==3.3.4
atomicwrites==1.4.1
attrs==22.1.0
beautifulsoup4==4.12.2
cachetools==4.2.2
certifi==2022.12.7
cffi==1.16.0
chardet==3.0.4
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.4
cubes==1.0a3
defusedxml==0.8.0rc2
Django==3.2
django-colorfield==0.10.1
django-cors-headers==3.13.0
django-debug-toolbar==2.2
django-filter==23.3
django-rest-framework==0.1.0
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-extra-fields==3.7.0
drf-yasg==1.21.7
exceptiongroup==1.1.0
Faker==12.0.1
filetype==1.2.0
flake8==3.9.2
flake8-docstrings==1.6.0
Flask==2.2.2
fpdf==1.7.2
greenlet==2.0.1
idna==2.8
importlib-metadata==6.0.0
inflection==0.5.1
iniconfig==1.1.1
itsdangerous==2.1.2
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.17.3
MarkupSafe==2.1.1
mccabe==0.6.1
mixer==7.1.2
more-itertools==9.0.0
oauthlib==3.2.2
packaging==21.3
Pillow==10.1.0
pluggy==0.13.1
py==1.11.0
pycodestyle==2.7.0
pycparser==2.21
pydocstyle==6.3.0
pyflakes==2.3.1
PyJWT==2.1.0
pyparsing==3.0.9
pyrsistent==0.19.3
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
python-dateutil==2.8.2
python-dotenv==0.19.0
python-telegram-bot==13.7
python3-openid==3.2.0
pytz==2021.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0.1
reportlab==4.0.6
requests==2.26.0
requests-oauthlib==1.3.1
six==1.16.0
snowballstemmer==2.2.0
social-auth-app-django==4.0.0
social-auth-core==4.4.2
sorl-thumbnail==12.7.0
soupsieve==2.4
SQLAlchemy==1.4.46
sqlparse==0.4.1
toml==0.10.2
tomli==2.0.1
tornado==6.2
tzdata==2022.7
tzlocal==4.3
uritemplate==4.1.1
urllib3==1.25.11
wcwidth==0.2.6
webcolors==1.13
Werkzeug==2.2.2
zipp==3.11.0
psycopg2>=2.8,<3.0